log additional user-agent information
when a request comes from a web browser, the User-Agent in the request
header is the user's web browser, but what's more important in this
situation is which javascript program (ie web site) the browser had
open. the relevant information is always encoded in the Origin header
(for all modern browser's requests), and some clients set their own
version etc in the X-Client header.